+++ utils.py
@@ -1,0 +1,7 @@
+from typing import Any
+import yaml
+
+
+def get_config(filename: str = "config.yaml") -> dict[str, Any]:
+    with open(filename, encoding="utf-8") as file:
+        return yaml.safe_load(file)
